Proposes Code of Conduct
This is one of the two options Github provides based on the following project: https://www.contributor-covenant.org/faq/
We could also use the OSGeo one: http://www.osgeo.org/code_of_conduct/

The suggested contact email address is `hello@opertusmundi.eu`. Not sure if this is an appropriate one.